Fix Chrome supports query and get rid of sticky footer
Sticky footer causes overflow/scrollbar issues in FF. Since it kinda sucks in general, we'll just eliminate it instead of fixing it. If you wanna stick a credit somewhere else, that's fine, but I'm not worried about it either way.

I borked the Chromium supports query the other day. The new one should do the trick.